requirements.txt
connexion[swagger-ui]
aiohttp_jinja2
peewee
